schedule_reader_function/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
protobuf==3.20.1 